toolkit/components/extensions/test/mochitest/test_ext_background_generated_url.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test _generated_background_page.html</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
</head>
<body>

<script>
"use strict";

add_task(function* test_url_of_generated_background_page() {
  function backgroundScript() {
    const EXPECTED_URL = browser.runtime.getURL("/_generated_background_page.html");
    browser.test.assertEq(EXPECTED_URL, location.href);
    browser.test.sendMessage("script done", EXPECTED_URL);
  }
  let extension = ExtensionTestUtils.loadExtension({
    manifest: {
      background: {
        scripts: ["bg.js"],
      },
      web_accessible_resources: ["_generated_background_page.html"],
    },
    files: {
      "bg.js": backgroundScript,
    },
  });

  yield extension.startup();
  const EXPECTED_URL = yield extension.awaitMessage("script done");

  let win = window.open(EXPECTED_URL);
  ok(win, "Should open new tab at URL: " + EXPECTED_URL);
  yield extension.awaitMessage("script done");
  win.close();

  yield extension.unload();
});

</script>
</body>
</html>
